+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- `cargo-nextest` installation no longer use `cargo-binstall`.
+  This improves security, performance, robustness of installation. See [#487](https://github.com/taiki-e/install-action/issues/487) for more.
+
 ## [2.37.0] - 2024-06-08
 
 - Support `rclone` ([#511](https://github.com/taiki-e/install-action/pull/511), thanks @jayvdb)